implemented initial CRUD methods in the service

+++ src/main/java/timekeeper/users/repositories/UserRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import timekeeper.users.models.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {}
+public interface UserRepository extends JpaRepository<User, Long> {}
